+++ CHANGELOG.rst
@@ -1,8 +1,8 @@
 Unreleased changes
 ------------------
 
-* tohdf5.py: adds a hit time conversion option from JTE to MC time.
-
+* tohdf5.py: - adds a hit time conversion option from JTE to MC time.
+             - adds a mc_tracks time conversion option from MC to JTE time.
 
 Version 8
 ---------